+++ .gitignore
@@ -1,2 +1,2 @@
-venv
+.venv
 .idea
