--- app/views.py
@@ -1,12 +1,0 @@
-from flask import render_template
-from app import app
-
-
-@app.route('/')
-def index():
-    return render_template("index.html")
-
-
-@app.route('/about')
-def about():
-    return render_template("about.html")
